Set up pytest testing with a simple test.
