+++ requirements-test.txt
@@ -1,0 +1,1 @@
+pytest==7.4.3
